backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
motor>=3.3.0
beanie>=1.25.0
certifi
websockets
google-genai
numpy
ffmpeg-python